+++ R/aaa.R
@@ -1,0 +1,3 @@
+#' Internal package environment for usmapdata.
+#' @keywords internal
+.pkg_env <- new.env(parent = emptyenv())